worker/requirements.txt
supabase
anthropic
youtube-transcript-api
google-api-python-client
requests
python-dotenv
